refactor(integrations): give eight more vendors a public API (setup, client)
