+++ surfaces/cli/wizard/configurators/observability.py
@@ -4,12 +4,12 @@ from __future__ import annotations
 
 from config.env_file import sync_env_values
 from infrastructure.terminal.theme import ERROR, HIGHLIGHT, SECONDARY
-from integrations.coralogix.setup import CORALOGIX_SETUP
+from integrations.coralogix import CORALOGIX_SETUP
 from integrations.datadog.setup import DATADOG_SETUP
 from integrations.grafana.setup import GRAFANA_SETUP
-from integrations.honeycomb.setup import HONEYCOMB_SETUP
-from integrations.new_relic.setup import NEW_RELIC_SETUP
-from integrations.opensearch.setup import OPENSEARCH_SETUP
+from integrations.honeycomb import HONEYCOMB_SETUP
+from integrations.new_relic import NEW_RELIC_SETUP
+from integrations.opensearch import OPENSEARCH_SETUP
 from integrations.store import remove_integration, upsert_integration
 from integrations.tempo.setup import TEMPO_SETUP
 from surfaces.cli.wizard._ui import (
